Refactor aggregation of data for batchmaker views
- Uses an issuequeue.Queue in batchmakerhandler to split up *actually*
  ready-for-batching issues from those still under embargo
- Changes the pluralize method back to use an int instead of int64 to
  fix other uses of it
- Adds an easy way to get number of issues from an issuequeue.Queue

+++ src/cmd/server/internal/responder/templates.go
@@ -79,7 +79,7 @@ func actionVerb(at string) string {
 	return models.ActionType(at).Describe()
 }
 
-func pluralize(singular, plural string, count int64) string {
+func pluralize(singular, plural string, count int) string {
 	if count == 1 {
 		return fmt.Sprintf("%d %s", count, singular)
 	}
